requirements.txt
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-benchmark>=4.0.0
tqdm>=4.65.0
black>=23.0.0
ruff>=0.0.280
pydantic>=2.0.0
pyyaml>=6.0.0
nvidia-ml-py3>=7.352.0
httpx>=0.24.0

